fixing EVEEERYYYTHIIIING

+++ postDetails.php
@@ -2,13 +2,16 @@
     use imdmedia\Feed\Post;
     use imdmedia\Auth\Security;
     use imdmedia\Auth\User;
-use imdmedia\Feed\ReportPost;
+    use imdmedia\Feed\ReportPost;
+    use imdmedia\Feed\Comment;
+    use imdmedia\Feed\Like;
 
     require __DIR__ . '/vendor/autoload.php';
-    include_once("inc/functions.inc.php");
-    
-    boot();
+
+    session_start();
     $auth = Security::checkLoggedIn();
+    Security::onlyLoggedInUsers();
+
 
     //add search function
     if(isset($_GET['search'])){
@@ -16,39 +19,72 @@ use imdmedia\Feed\ReportPost;
     }
 
     if(isset($_GET['Post'])){
-        $postId = $_GET['Post'];
-        $postDetails = Post::getPostById($postId);
-        $posts = Post::getPostByTags($postDetails['tags']);
-        $reportCheck = ReportPost::reportCheck($postId ,$_SESSION['user']['id']);
-        if($postDetails['userid'] == $_SESSION['user']['id']){
-            $ownProfile = true;
+        try { 
+            $postId = $_GET['Post'];
+            $postDetails = Post::getPostById($postId);
+            $posts = Post::getDetailPostsByTags($postDetails['tags'], $postId);
+            $reportCheck = ReportPost::reportCheck($postId ,$_SESSION['user']['id']);
+            if($postDetails['userid'] == $_SESSION['user']['id']){
+                $ownProfile = true;
+            }
+            else $ownProfile = false;
+        } catch (Throwable $e) {
+            $error = $e->getMessage();
         }
     }
+    else (
+        header("Location: index.php")
+    );
 
-    if ($_SERVER['REQUEST_METHOD'] == 'POST') {
-        if ($_POST['confirm'] == 'Yes') {
-            Post::deletePostById($postId);
-            header("Location: index.php");
-        }
-    }
+   if (isset($_POST['confirm'])) {
+       if ($_SERVER['REQUEST_METHOD'] == 'POST') {
+           try {
+               if ($_POST['confirm'] == 'Yes') {
+                   Post::deletePostById($postId);
+                   header("Location: index.php");
+               }
+           } catch (Throwable $e) {
+               $error = $e->getMessage();
+           }
+       }
+   }
 
     if(isset($_POST['changeTitle'])){
-        $newTitle = $_POST["newTitle"];
-        $postId = $_GET['Post'];
-        $post = new Post();
-        $changeTitle = $post->changeTital($newTitle, $postId);
-
+        try {
+            $newTitle = $_POST["newTitle"];
+            $postId = $_GET['Post'];
+            $post = new Post();
+            $changeTitle = $post->changeTitle($newTitle, $postId);
+        } catch (Throwable $e) {
+            $error = $e->getMessage();
+        }
 
     }
 
     if(isset($_POST['changeTag'])){
-        $newTag = $_POST["newTag"];
-        $postId = $_GET['Post'];
-        $post = new Post();
-        $changeTag = $post->changeTag($newTag, $postId);
-
+        try {
+            $newTag = $_POST["newTag"];
+            $postId = $_GET['Post'];
+            $post = new Post();
+            $post->setTags($_POST["newTag"]);
+            $changeTag = $post->changeTag($postId);
+        }   catch (Throwable $e) {
+            $error = $e->getMessage();
+        } 
 
     }
+    $like = "Like";
+    if(isset($_SESSION['user']['id'])){
+            $checklike = Like::checkLiked($_SESSION['user']['id'], $postId);
+             if($checklike == 1) {
+                $like = "Unlike";
+            }
+            else {
+                $like = "Like";
+            }
+    }
+    
+    $totalLikes = Like::getAll($postId);  
 
 
 
@@ -56,15 +92,61 @@ use imdmedia\Feed\ReportPost;
 <html lang="en">
 <head>
     <?php include_once("inc/header.inc.php"); ?>
-    <title><?php echo $postDetails['title']; ?></title>
+    <title><?php echo htmlspecialchars($postDetails['title']); ?></title>
 </head>
 <body>
     <?php include_once("inc/nav.inc.php"); ?>
 
-    <section class="details">
-        <div class="details__img">
-            <img class="details__imgSRC" src="<?php echo $postDetails['filePath']; ?>" alt="<?php echo $postDetails['title']; ?>">
-        </div>
+<div class="detailsblock">
+    <section class="postL">
+        <section class="postImage">
+            <div class="details__img">
+                <img class="details__imgSRC" src="<?php echo $postDetails['filePath']; ?>" alt="<?php echo htmlspecialchars($postDetails['title']); ?>">
+            </div>
+        </section>
+        <section class="feed profileFeed">
+            <?php foreach ($posts as $key => $post): ?>
+            <?php    
+                
+                $feedlike = "Like";
+                    if(isset($_SESSION['user']['id'])){
+                        $feedchecklike = Like::checkLiked($_SESSION['user']['id'], $post[('id')]);
+                    if($feedchecklike == 1) {
+                        $feedlike = "Unlike";
+                     }
+                    else {
+                        $feedlike = "Like";
+                    }
+                }
+    
+                $feedtotalLikes = Like::getAll($post[('id')]);   
+            ?>    
+            <div class="post">
+                <a href="postDetails.php?Post=<?php echo htmlspecialchars($post['id']); ?>">
+                <img src="<?php echo $post['filePath']; ?>" alt="<?php echo htmlspecialchars($post['title']); ?>"></a>
+                <h3><?php echo htmlspecialchars($post['title']); ?></h3>
+                <?php if ($auth == true): ?>
+                <a href="account.php?Account=<?php echo htmlspecialchars($post['userName']); ?>" class="postUsername"><?php echo htmlspecialchars($post['userName']); ?></a>
+                <?php $tags = json_decode($post['tags']); ?>
+                <?php foreach ($tags as $tag): ?>
+                <a href="index.php?search=<?php echo htmlspecialchars($tag); ?>" class="text__tags">#<?php echo htmlspecialchars($tag); ?></a>
+                <?php endforeach ?>
+                <?php endif ?>
+                <p class="likes">
+                        <a style="text-decoration: none;" href="#" 
+                        id="likeButton-<?php echo htmlspecialchars($post[('id')]); ?>"
+                        onclick="likePost(this, <?php echo htmlspecialchars($post[('id')]); ?>)" 
+                        >
+                        <?php echo $feedlike;?>
+                        </a>
+                            <span id="totalLikes-<?php echo htmlspecialchars($post[('id')]); ?>"><?php echo $feedtotalLikes;?></span>
+                        </p>
+                </p>
+            </div>
+            <?php endforeach; ?>        
+        </section>
+    </section>
+    <section class="postDetails">
         <div class="details__details">
             <div class="details__header">
                 <div>
@@ -87,10 +169,10 @@ use imdmedia\Feed\ReportPost;
                             <?php endif; ?>
                             <?php if($postDetails['userName'] !== $_SESSION['user']['username']): ?>
                                 <?php if ($reportCheck): ?>
-                                    <button class="report" data-check="reported" data-post="<?php echo $postId; ?>" data-id="<?php echo $_SESSION['user']['id'];?>">Undo report</button>
+                                    <button class="report" data-check="reported" data-post="<?php echo htmlspecialchars($postId); ?>" data-id="<?php echo $_SESSION['user']['id'];?>">Undo report</button>
                                     <input class="session" type="hidden" value="">
                                 <?php else: ?>
-                                    <button class="report" data-check="report" data-post="<?php echo $postId; ?>" data-id="<?php echo $_SESSION['user']['id'];?>">Report this</button>
+                                    <button class="report" data-check="report" data-post="<?php echo htmlspecialchars($postId); ?>" data-id="<?php echo $_SESSION['user']['id'];?>">Report this</button>
                                     <input class="session" type="hidden" value="">
                                 <?php endif; ?>
                             <?php endif; ?>
@@ -104,98 +186,105 @@ use imdmedia\Feed\ReportPost;
             <div class="details__text details__description">
                 <p class="details__text details__text__subtitle details__text__subtitle--description">Description</p>
                 <?php if(isset($postDetails['description'])): ?>
-                    <p class="details__text details__text--description"><?php echo $postDetails['description']; ?> </p>
+                    <p class="details__text details__text--description"><?php echo htmlspecialchars($postDetails['description']); ?> </p>
                 <?php elseif(!isset($postDetails['description'])): ?>
                     <p class="details__text details__text--description" >No description given</p>
                 <?php endif; ?>
             </div>
 
             <?php if ($auth == true): ?>
-            <?php foreach ($tags as $tag): ?>
-            <a href="?tags=<?php echo htmlspecialchars($tag); ?>" class="details__text text__tags details__text--tags">#<?php echo htmlspecialchars($tag); ?></a>
-            <?php endforeach ?>
-            <?php if(isset($ownProfile)): ?>
-            <button class="deleteBtn">Delete</button>
-            <form class="delete" action="" method="post">
-                <div class="confirmation">
-                <p>Are you sure?</p>
-                <div class="deleteBtns">
-                    <input class="yesBtn confirmationBtn" type="submit" name="confirm" value="Yes">
-                    <input class="noBtn confirmationBtn" type="submit" name="confirm" value="No">
+                <div class="postData">
+                    <div>
+                    <a style="text-decoration: none;" href="#" 
+                        id="likeButton-<?php echo htmlspecialchars($postId); ?>"
+                        onclick="likePost(this, <?php echo htmlspecialchars($postId); ?>)">
+                        <?php echo $like;?>
+                        </a>
+                        <p id="totalLikes-<?php echo htmlspecialchars($postId); ?>"><?php echo $totalLikes;?></p>
+                    </div>
+                    <?php foreach ($tags as $tag): ?>
+                        <a href="index.php?search=<?php echo htmlspecialchars($tag); ?>" class="details__text text__tags details__text--tags">#<?php echo htmlspecialchars($tag); ?></a>
+                    <?php endforeach ?>
                 </div>
+                <?php if($ownProfile == true): ?>
+                    <form class="delete" action="" method="post">
+                        <div class="confirmation">
+                        <p>Are you sure?</p>
+                        <div class="deleteBtns">
+                            <input class="yesBtn confirmationBtn" type="submit" name="confirm" value="Yes">
+                            <input class="noBtn confirmationBtn" type="submit" name="confirm" value="No">
+                        </div>
+                        </div>
+                    </form>
+                    <?php if ($reportCheck): ?>
+                        <p class="report" data-check="reported" data-post="<?php echo htmlspecialchars($postId); ?>" data-id="<?php echo $_SESSION['user']['id'];?>">You reported this!</p>
+                        <input class="session" type="hidden" value="">
+                    <?php else: ?>
+                        <p class="report" data-check="report" data-post="<?php echo htmlspecialchars($postId); ?>" data-id="<?php echo $_SESSION['user']['id'];?>">Report this</p>
+                        <input class="session" type="hidden" value="">
+                    <?php endif; ?>
+                <?php endif; ?>
+                <?php if($ownProfile == true): ?>
+                    <div class="changePostInfo">  
+                    <form class="postForm" method="POST" action="" enctype="multipart/form-data"> 
+                        <h2>Change title</h2>
+                        <?php if (isset($error)) : ?>
+                        <div class="error">
+                            <p>
+                                <?php echo $error; ?>
+                            </p>
+                        </div>
+                        <?php endif; ?>	
+                        <label for="pass">New Title</label>
+                        <input type="title" id="newTitle" name="newTitle">
+
+                        <button type="submit" name="changeTitle" class="upload__text__btn">Change Title</button>
+                    </form>  
+                    
+                    <form id="changeTag" class="postForm" method="POST" action="" enctype="multipart/form-data"> 
+                        <h2>Change tag</h2>
+                        <?php if (isset($error)) : ?>
+                        <div class="error">
+                            <p>
+                                <?php echo $error; ?>
+                            </p>
+                        </div>
+                        <?php endif; ?>	
+                        <label for="pass">New Tags</label>
+                        <input type="tag" id="newTag" name="newTag">
+
+                        <button type="submit" name="changeTag" class="upload__text__btn">Change Tags</button>
+                    </form> 
+                    </div>  
+                <?php endif; ?>
+                <div class="post__comments">
+                    <div class="post__comments__form">
+                        <input type="text" id="commentText" placeholder="Write a comment...">
+                        <a href="#" class="upload__text__btn" id="btnAddComment"
+                        data-post-Id="<?php echo htmlspecialchars($postId); ?>"
+                        >Add comment</a>
+                    </div>
+                    <?php $comments = Comment::getAll($postId); ?>
+                    <ul class="post__comments__list">
+                    <?php foreach($comments as $comment): ?>
+                    <?php $user = User::getUserById($comment['userId']); ?>    
+                        <li class="comment">
+                            <p><?php echo htmlspecialchars($user['username']);?></p>
+                            <p><?php echo htmlspecialchars($comment['comment']);?></p>
+                        </li>        
+                    <?php endforeach; ?>    
+                    </ul>
+
                 </div>
-            </form>
-            <?php if ($reportCheck): ?>
-                <p class="report" data-check="reported" data-post="<?php echo $postId; ?>" data-id="<?php echo $_SESSION['user']['id'];?>">You reported this!</p>
-                <input class="session" type="hidden" value="">
-            <?php else: ?>
-                <p class="report" data-check="report" data-post="<?php echo $postId; ?>" data-id="<?php echo $_SESSION['user']['id'];?>">Report this</p>
-                <input class="session" type="hidden" value="">
-            <?php endif; ?>
-            <?php endif; ?>
             <?php endif ?>
         </div>
     </section>
-
-    <section class="feed profileFeed">
-        <?php foreach ($posts as $key => $post): ?>
-        <div class="post">
-            <img src="<?php echo $post['filePath']; ?>" alt="<?php echo $post['title']; ?>">
-            <h3><?php echo htmlspecialchars($post['title']); ?></h3>
-            <?php if ($auth == true): ?>
-            <a href="account.php?Account=<?php echo htmlspecialchars($post['userName']); ?>" class="postUsername"><?php echo htmlspecialchars($post['userName']); ?></a>
-            <?php $tags = json_decode($post['tags']); ?>
-            <?php foreach ($tags as $tag): ?>
-            <a href="?tags=<?php echo htmlspecialchars($tag); ?>" class="postTags">#<?php echo htmlspecialchars($tag); ?></a>
-            <?php endforeach ?>
-            <?php endif ?>
-            <p class="likes">
-                            <a href="#" id="btnAddLike">&hearts;</a>
-                            <span id="likes_counter">x</span> people like this
-                        </p>
-
-            <div class="post__comments">
-                <div class="post__comments__form">
-                    <input type="text" id="commentText" style="color:black">
-                    <a href="#" class="btn" id="btnAddComment" data-postId="<?php echo $_GET['post']?>">Add comment</a>
-                </div>  
-                
-                <ul class="post__comments__list">
-                    <?php foreach($allComments as $c): ?>
-                    <li><?php echo $c['message']; ?></li>  
-                    <?php endforeach; ?>
-                </ul>
-            </div>   
-        </div>
-        <?php endforeach; ?>
-         
-        
-        <form class="postForm" method="POST" action="" enctype="multipart/form-data"> 
-            <h2>Change title</h2>
-            <div class="error">
-                <?php echo $error ?>
-            </div>
-            <label for="pass">New Title</label>
-            <input type="title" id="newTitle" name="newTitle">
-
-            <button type="submit" name="changeTitle">Change Title</button>
-        </form>  
-        
-        <form class="postForm" method="POST" action="" enctype="multipart/form-data"> 
-            <h2>Change tag</h2>
-            <div class="error">
-                <?php echo $error ?>
-            </div>
-            <label for="pass">New Tags</label>
-            <input type="tag" id="newTag" name="newTag">
-
-            <button type="submit" name="changeTag">Change Tags</button>
-        </form> 
-        
-    </section>
+</div>
 
     <script type="module" src="./js/sass.js"></script>
     <script src="js/deletePost.js"></script>
     <script src="js/postDetails.js"></script>
+    <script src="js/comment.js"></script>
+    <script src="js/like.js"></script>
 </body>
 </html>
